pyproject.toml
[tool.poetry]
name = "grpc-interceptor"
version = "0.11.0"
description = "Simplifies gRPC interceptors"
license = "MIT"
readme = "README.md"
homepage = "https://github.com/d5h-foss/grpc-interceptor"
repository = "https://github.com/d5h-foss/grpc-interceptor"
keywords = ["grpc", "interceptor"]
authors = ["Dan Hipschman"]
documentation = "https://grpc-interceptor.readthedocs.io"

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.8.0"
protobuf = {version = ">=3.6.0", optional = true}

[tool.poetry.extras]
testing = ["protobuf"]

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
# Use an old version of protoc, because the code generated by newer versions sometimes
# require newer versions of the protobuf package as well.
grpcio-tools = "1.8.1"
coverage = {extras = ["toml"], version = "^5.2"}
pytest-cov = "^2.10.0"
black = "^19.10b0"
flake8 = "^3.8.3"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.9.0"
mypy = "^0.782"
mypy-protobuf = "^1.23"
flake8-docstrings = "^1.5.0"
sphinx = "^3.1.2"
codecov = "^2.1.8"
xdoctest = "^0.13.0"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["grpc_interceptor"]
omit = ["*_pb2.py", "*_pb2_grpc.py"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
